# Optimize AssetRegistry.find and findAll