Improves `AssetRegistry.find` and `findAll` performance, by building and managing internal index of assets by names. And avoids casting large asset lists to arrays during these methods to avoid GC stalls.

In a project with around 2.5k of assets, calling `find` was taking around 0.04ms per request. With this patch it takes around 0.00006ms. This is around ~666 times improvement.
`findAll` used to take 0.04ms per request also, now it takes 0.0006ms. ~66 times improvement.

**In specific project, during custom scene loaders that reference assets by names, it used to take around 200ms just finding assets, now it takes 0.25ms.**

_Timings are just a relative measurement and based on hardware._
_It only highlights the performance change between versions._
_Tested on high-end computer, so on mid/low/mobile it is even more apparent improvement._

I confirm I have read the [contributing guidelines](https://github.com/playcanvas/engine/blob/master/.github/CONTRIBUTING.md) and signed the [Contributor License Agreement](https://docs.google.com/a/playcanvas.com/forms/d/1Ih69zQfJG-QDLIEpHr6CsaAs6fPORNOVnMv5nuo0cjk/viewform).